requirements.txt
celery==5.3.6
redis==5.0.3
playwright==1.52.0
anthropic==0.25.0
sqlalchemy==2.0.29
alembic==1.13.1
fastapi==0.111.0
uvicorn==0.29.0
python-dotenv==1.0.1
httpx==0.27.0
jinja2==3.1.3
pillow==10.4.0
requests==2.31.0
pytest==8.1.1
pytest-asyncio==0.23.6
